requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
anyio==4.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asn1crypto==1.5.1
asttokens==2.4.1
async-lru==2.0.4
attrs==23.1.0
Babel==2.13.1
beautifulsoup4==4.12.2
bertopic==0.16.2
bleach==6.1.0
boto3==1.38.14
botocore==1.38.14
certifi==2023.7.22
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.0
contourpy==1.2.0
cryptography==44.0.3
cycler==0.12.1
Cython==0.29.37
datasets==2.19.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.8
et_xmlfile==2.0.0
executing==2.0.1
fastjsonschema==2.18.1
filelock==3.13.4
fonttools==4.50.0
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.3.1
graphviz==0.20.3
hdbscan==0.8.36
hiddenlayer==0.3
huggingface-hub==0.23.0
idna==2.10
intel-openmp==2021.4.0
ipykernel==6.26.0
ipython==8.17.2
ipywidgets==8.1.3
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter-events==0.9.0
jupyter-lsp==2.2.0
jupyter_client==8.6.0
jupyter_core==5.5.0
jupyter_server==2.10.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.8
jupyterlab-pygments==0.2.2
jupyterlab_server==2.25.1
jupyterlab_widgets==3.0.11
kiwisolver==1.4.5
llvmlite==0.42.0
MarkupSafe==2.1.3
matplotlib==3.8.3
matplotlib-inline==0.1.6
missingno==0.5.2
mistune==3.0.2
mkl==2021.4.0
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
nbclient==0.9.0
nbconvert==7.11.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.3
nltk==3.8.1
nose==1.3.7
notebook==7.0.6
notebook_shim==0.2.3
numba==0.59.1
numpy==1.26.2
openpyxl==3.1.5
overrides==7.4.0
packaging==23.2
pandas==2.1.4
pandocfilters==1.5.0
parso==0.8.3
pillow==10.3.0
platformdirs==4.0.0
plotly==5.22.0
portalocker==2.8.2
prettytable==3.9.0
prometheus-client==0.18.0
prompt-toolkit==3.0.40
psutil==5.9.6
pure-eval==0.2.2
pyarrow==16.1.0
pyarrow-hotfix==0.6
pycparser==2.21
pyecharts==2.0.4
Pygments==2.16.1
PyJWT==2.10.1
pynndescent==0.5.12
pyOpenSSL==25.0.0
pyparsing==3.1.2
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3.post1
pywin32==306
pywinpty==2.0.12
PyYAML==6.0.1
pyzmq==25.1.1
referencing==0.30.2
regex==2024.5.15
requests==2.25.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.12.0
s3transfer==0.12.0
safetensors==0.4.3
scikit-learn==1.4.1.post1
scipy==1.12.0
seaborn==0.13.2
Send2Trash==1.8.2
sentence-transformers==3.0.0
setuptools==70.0.0
simplejson==3.19.2
six==1.16.0
skorch==0.15.0
sniffio==1.3.0
snowflake-connector-python==3.15.0
sortedcontainers==2.4.0
soupsieve==2.5
stack-data==0.6.3
sympy==1.12
tabulate==0.9.0
tbb==2021.12.0
tenacity==8.3.0
terminado==0.18.0
threadpoolctl==3.4.0
tinycss2==1.2.1
tokenizers==0.19.1
tomlkit==0.13.2
torch==2.2.2
torchattacks==3.5.1
torchaudio==2.2.2
torchdata==0.7.1
torchinfo==1.8.0
torchtext==0.17.2
torchview==0.2.6
torchvision==0.17.2
tornado==6.3.3
tqdm==4.66.2
traitlets==5.13.0
transformers==4.41.2
types-python-dateutil==2.8.19.14
typing_extensions==4.11.0
tzdata==2023.4
umap-learn==0.5.6
uri-template==1.3.0
urllib3==1.26.20
vaderSentiment==3.3.2
wcwidth==0.2.9
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
widgetsnbextension==4.0.11
xxhash==3.4.1
yarl==1.9.4
